Platforms 1
Moved player control and movement and moving platforms to Update rather
than FixedUpdate as we're not using Rigidbodies and the physics engine.
Added phase-able platforms and a countdown platform (destroys itself a
set time after the player steps on it).
Added the RayLayers static class to hold info on LayerMasks the player
will interact with.
Made the skin variable in PlayerPhysics static to allow other classes to
access it without created a reference of the object.